Move all backend-specific code into the ops package.
