+++ .github/workflows/build.yaml
@@ -81,10 +81,10 @@ jobs:
             | tee artifacts/tests/output.txt
       - name: Upload Test Results
         uses: actions/upload-artifact@v2
+        if: ${{ always() }}
         with:
           name: test-results
           path: artifacts/tests
-        if: ${{ always() }}
       - name: Report Test Results
         uses: EnricoMi/publish-unit-test-result-action/composite@v1
         if: ${{ success() && github.event_name == 'pull_request_target'}}
@@ -92,7 +92,7 @@ jobs:
           github_token: ${{github.token}}
           files: ./artifacts/tests/results.xml
       - name: Report Coverage
-        uses: MishaKav/pytest-coverage-comment@v1.1.6
+        uses: MishaKav/pytest-coverage-comment@main
         if: ${{ success() && github.event_name == 'pull_request_target'}}
         with:
           github-token: ${{github.token}}
