+++ clients/context.py
@@ -1,7 +1,3 @@
-import sys
-sys.path.insert(0, "..")
-
-
 from sameproject.data.step import Step
 from sameproject.ops import notebooks
-from sameproject.backends.durable_functions.constants import EXECUTE_WORKFLOW_ACTIVITY_NAME
+from sameproject.ops.durable_functions.constants import EXECUTE_WORKFLOW_ACTIVITY_NAME
