--- sameproject/__init__.py
@@ -1,5 +1,0 @@
-"""The command line entry point for SAME"""
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parent.parent))
