--- sameproject/backends/durable_functions/.vscode/extensions.json
@@ -1,6 +1,0 @@
-{
-  "recommendations": [
-    "ms-azuretools.vscode-azurefunctions",
-    "ms-python.python"
-  ]
-}
