--- sameproject/backends/durable_functions/.vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "azureFunctions.deploySubpath": ".",
-    "azureFunctions.scmDoBuildDuringDeployment": true,
-    "azureFunctions.projectLanguage": "Python",
-    "azureFunctions.projectRuntime": "~3",
-    "debug.internalConsoleOptions": "neverOpen"
-}
