--- sameproject/backends/durable_functions/.vscode/tasks.json
@@ -1,26 +1,0 @@
-{
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "type": "func",
-      "command": "host start --verbose",
-      "problemMatcher": "$func-python-watch",
-      "isBackground": true,
-      "dependsOn": "pip install (functions)"
-    },
-    {
-      "label": "pip install (functions)",
-      "type": "shell",
-      "osx": {
-        "command": "python -m pip install -r requirements.txt"
-      },
-      "windows": {
-        "command": "python -m pip install -r requirements.txt"
-      },
-      "linux": {
-        "command": "python -m pip install -r requirements.txt"
-      },
-      "problemMatcher": []
-    }
-  ]
-}
