+++ sameproject/ops/aml/__init__.py
@@ -1,0 +1,2 @@
+from .render import render
+from .deploy import deploy
