+++ sameproject/ops/durable_functions/.funcignore
