+++ sameproject/ops/durable_functions/common/http_utils.py
