+++ sameproject/ops/kubeflow/render.py
@@ -1,37 +1,50 @@
-from sameproject.data.step import Step
-from pathlib import Path
-from sameproject.ops import helpers
 from jinja2 import Environment, FileSystemLoader, select_autoescape
-
+from kubernetes import client, config
+from base64 import urlsafe_b64encode
+from pathlib import Path
 from typing import Tuple
-
-import logging
-
 from uuid import uuid4
+import logging
+import dill
+import os
+
+from sameproject.data.step import Step
+from sameproject.ops import helpers
 
 
-kubeflow_root_template = "aml/root.jinja"
-kubeflow_step_template = "aml/step.jinja"
+kubeflow_run_info_template = "run_info.jinja"
+kubeflow_root_template = "root.jinja"
+kubeflow_step_template = "step.jinja"
 
 
-def render_function(compile_path: str, steps: list, same_config: dict) -> Tuple[Path, str]:
+def render(compile_path: str, steps: list, same_config: dict) -> Tuple[Path, str]:
     """Renders the notebook into a root file and a series of step files according to the target requirements. Returns an absolute path to the root file for deployment."""
-    templateLoader = FileSystemLoader(searchpath="./templates")
-    env = Environment(loader=templateLoader)
-    same_config["compile_path"] = compile_path
+    templateDir = os.path.dirname(os.path.abspath(__file__))
+    templateLoader = FileSystemLoader(templateDir)
+    print(f"Template dir {templateDir}")
+    env = Environment(trim_blocks=True, loader=templateLoader)
+
+    # Write the steps first so that if we need to make any changes while writing (such as adding a unique name), it's reflected in the root filepath
+    for step_name in steps:
+        step_file_string = _build_step_file(env, steps[step_name])
+        helpers.write_file(Path(compile_path) / f"{steps[step_name].unique_step_name }.py", step_file_string)
+
     root_file_string = _build_root_file(env, steps, same_config)
 
     root_pipeline_name = f"root_pipeline_{uuid4().hex.lower()}"
     root_path = Path(compile_path) / f"{root_pipeline_name}.py"
     helpers.write_file(root_path, root_file_string)
 
-    for step_name in steps:
-        # Need a unique name so that libraries don't conflict in sys.modules. This is MOSTLY a test issue, but could be the case generally.
-        step_file_string = _build_step_file(env, steps[step_name], steps[step_name].unique_step_name)
-        (Path(compile_path) / steps[step_name].unique_step_name).mkdir()
-        helpers.write_file(Path(compile_path) / steps[step_name].unique_step_name / f"{steps[step_name].unique_step_name}.py", step_file_string)
+    run_info_path = Path(compile_path) / "run_info.py"
+    run_info_file_string = _build_run_info_file(env)
+    helpers.write_file(run_info_path, run_info_file_string)
 
     return (compile_path, root_pipeline_name)
+
+
+def _build_run_info_file(env: Environment) -> str:
+    template = env.get_template(kubeflow_run_info_template)
+    return template.render()
 
 
 def _build_root_file(env: Environment, all_steps: list, same_config: dict) -> str:
@@ -47,8 +60,6 @@ def _build_root_file(env: Environment, all_steps: list, same_config: dict) -> st
         "experiment_name_safe": "",
         "list_of_environments": {},
         "image_pull_secrets": {},
-        "aml_workspace_credentials": {},
-        "compile_dir": "",
     }
 
     params_to_merge = []
@@ -85,8 +96,6 @@ def _build_root_file(env: Environment, all_steps: list, same_config: dict) -> st
 
         if root_contract["list_of_environments"][name]["private_registry"]:
 
-            # This is starting to have quite a lot of code smell - root_contract requires a bit of massaging (instead of
-            # just passing through same_config to the jinja template nakedly) but i'm starting to dislike everything here.
             if "credentials" in same_config.environments[name]:
                 # Someone COULD set this to be a 'private_registry' but did not set credentials. This may be ok!
                 # They could have already mounted the secret in the cluster, so we should let it go ahead.
@@ -102,19 +111,7 @@ def _build_root_file(env: Environment, all_steps: list, same_config: dict) -> st
                 these_credentials["image_pull_secret_username"] = same_config.environments[name].credentials.get("image_pull_secret_username", "")
                 these_credentials["image_pull_secret_password"] = same_config.environments[name].credentials.get("image_pull_secret_password", "")
                 these_credentials["image_pull_secret_email"] = same_config.environments[name].credentials.get("image_pull_secret_email", "")
-
                 root_contract["secrets_to_create_as_dict"][name] = these_credentials
-
-    if same_config.get("aml"):
-        root_contract["aml_workspace_credentials"] = {
-            "AML_SP_PASSWORD_VALUE": same_config.aml.AML_SP_PASSWORD_VALUE,
-            "AML_SP_TENANT_ID": same_config.aml.AML_SP_TENANT_ID,
-            "AML_SP_APP_ID": same_config.aml.AML_SP_APP_ID,
-            "WORKSPACE_SUBSCRIPTION_ID": same_config.aml.WORKSPACE_SUBSCRIPTION_ID,
-            "WORKSPACE_RESOURCE_GROUP": same_config.aml.WORKSPACE_RESOURCE_GROUP,
-            "WORKSPACE_NAME": same_config.aml.WORKSPACE_NAME,
-            "AML_COMPUTE_NAME": same_config.aml.AML_COMPUTE_NAME,
-        }
 
     # Until we get smarter, we're just going to combine inject EVERY package into every step.
     # This is not IDEAL, but it's not as bad as it sounds because it'll allow systems to cache
@@ -175,23 +172,16 @@ def _build_root_file(env: Environment, all_steps: list, same_config: dict) -> st
 
     # List manipulation is also pretty weak in jinja (plus I like views being very non-functional). We'll
     # create the comma delim list of steps (which we need for DAG description) in python as well.
-
-    # For AML, each "step" needs to have '_step' attached (this may be historical)
-    # and not necessary - look at it when we combine all these step rendering functions into one
-    root_contract["comma_delim_list_of_step_names_as_str"] = ", ".join([f"{all_steps[this_step_name].unique_step_name}_step" for this_step_name in all_steps])
-
-    root_contract["compile_path"] = same_config["compile_path"]
+    root_contract["comma_delim_list_of_step_names_as_str"] = ", ".join([name for name in all_steps])
 
     return template.render(root_contract)
 
 
-def _build_step_file(env: Environment, step: Step, step_name: str) -> str:
-    template = env.get_template(kubeflow_step_template)
+def _build_step_file(env: Environment, step: Step) -> str:
+    step_contract = {
+        "name": step.name,
+        "unique_step_name": step.unique_step_name,
+        "inner_code": urlsafe_b64encode(dill.dumps(step.code)).decode()
+    }
 
-    # Create a parameter_string for putting in each step function
-    # default is to be a serialized empty dict. We should probably
-    # handle this a different way (allowing custom params to be passed in)
-    # but haven't found this requirement from a customer yet.
-    parameter_string = '__context="gAR9lC4=", __run_info="gAR9lC4=", __metadata_url=""'
-    step_contract = {"name": step_name, "inner_code": step.code, "parameter_string": parameter_string}
-    return template.render(step_contract)
+    return env.get_template(kubeflow_step_template).render(step_contract)
